+++ examples/actix_web_example/src/main.rs
@@ -49,7 +49,7 @@ async fn main() -> std::io::Result<()> {
         .idle_timeout(Some(std::time::Duration::from_secs(10 * 60)))
         .max_lifetime(Some(std::time::Duration::from_secs(30 * 60)))
         .reaper_rate(std::time::Duration::from_secs(5))
-        .min_idle(1)
+        .min_idle(0)
         .max_size(24)
         .build(mgr)
         .await